+++ src/googleTTS/GoogleTTS.ts
@@ -2,7 +2,8 @@ import fs from 'fs';
 import * as googleTTS from 'google-tts-api'; // ES6 or TypeScript
 import os from 'os';
 import path from 'path';
-import got from 'got';
+// import got from 'got';
+import axios from 'axios';
 import { GoogleTTSRequest } from './GoogleTTSRequest';
 import { CommonTTS } from '../common/commonTTS';
 
@@ -23,8 +24,10 @@ export class GoogleTTS extends CommonTTS<GoogleTTSRequest> {
                 slow: request.slow,
                 host: 'https://translate.google.com',
             });
-            const resp = await got.get(url, { encoding: 'binary' });
-            await fs.promises.writeFile(file, resp.rawBody);
+            // const resp = await got.get(url, { encoding: 'binary' });
+            // await fs.promises.writeFile(file, resp.rawBody);
+            const resp = await axios.get<Buffer>(url, { responseType: 'arraybuffer' });
+            await fs.promises.writeFile(file, resp.data);
             await super.log(request);
         }
         return file;